requirements_new.txt
asgiref==3.4.1
Babel==2.9.1
backports.zoneinfo==0.2.1
bleach==4.1.0
Django==4.0.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-api-logger==1.0.9
packaging==21.3
phonenumbers==8.12.41
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.6
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
